Remove Perfana configuration
