+++ modules/openapi-generator/src/main/resources/META-INF/services/org.openapitools.codegen.CodegenConfig
@@ -42,7 +42,7 @@ org.openapitools.codegen.languages.GraphQLSchemaCodegen
 org.openapitools.codegen.languages.GraphQLNodeJSExpressServerCodegen
 org.openapitools.codegen.languages.GroovyClientCodegen
 org.openapitools.codegen.languages.KotlinClientCodegen
-org.openapitools.codegen.languages.KotlinPerfanaGatlingCodegen
+org.openapitools.codegen.languages.KotlinGatlingCodegen
 org.openapitools.codegen.languages.KotlinServerCodegen
 org.openapitools.codegen.languages.KotlinSpringServerCodegen
 org.openapitools.codegen.languages.KotlinVertxServerCodegen
@@ -130,7 +130,7 @@ org.openapitools.codegen.languages.ScalaGatlingCodegen
 org.openapitools.codegen.languages.ScalaHttp4sClientCodegen
 org.openapitools.codegen.languages.ScalaHttp4sServerCodegen
 org.openapitools.codegen.languages.ScalaLagomServerCodegen
-org.openapitools.codegen.languages.ScalaPerfanaGatlingCodegen
+org.openapitools.codegen.languages.ScalaGatlingCodegen
 org.openapitools.codegen.languages.ScalaPlayFrameworkServerCodegen
 org.openapitools.codegen.languages.ScalaSttpClientCodegen
 org.openapitools.codegen.languages.ScalaSttp4ClientCodegen
