+++ modules/openapi-generator/src/main/resources/kotlin-gatling/build.gradle
@@ -1,0 +1,29 @@
+plugins {
+    id 'com.github.lkishalmi.gatling' version '0.7.1'
+}
+
+repositories {
+    maven { url "https://repo1.maven.org/maven2" }
+}
+
+dependencies {
+
+}
+
+apply plugin: "com.github.lkishalmi.gatling"
+
+gatling {
+    toolVersion = '2.3.0'
+    jvmArgs = ['-server', '-XX:+UseThreadPriorities',
+               '-XX:ThreadPriorityPolicy=42',
+               '-Xms2048M', '-Xmx2048M', '-Xmn500M',
+               '-XX:+HeapDumpOnOutOfMemoryError',
+               '-XX:+AggressiveOpts',
+               '-XX:+OptimizeStringConcat',
+               '-XX:+UseFastAccessorMethods',
+               '-XX:+UseParNewGC',
+               '-XX:+UseConcMarkSweepGC',
+               '-XX:+CMSParallelRemarkEnabled',
+               '-Djava.net.preferIPv4Stack=true',
+               '-Djava.net.preferIPv6Addresses=false']
+}
