+++ modules/openapi-generator/src/main/resources/kotlin-gatling/default.conf.mustache
@@ -1,0 +1,51 @@
+performance {
+    authorizationHeader = "~MANUAL_ENTRY~"
+    rampUpSeconds = 60
+    rampDownSeconds = 60
+    durationSeconds = 360
+    contentType = "application/json"
+    acceptType = "application/json"
+    rateMultiplier = 1
+    instanceMultiplier = 1
+    operationsPerSecond {
+    {{#apiInfo}}
+    {{#apis}}
+    {{#operations}}
+    {{#operation}}
+        {{operationId}} = 1
+    {{/operation}}
+    {{/operations}}
+    {{/apis}}
+    {{/apiInfo}}
+    }
+    global {
+        assertions {
+            responseTime {
+                min {
+                    lte = 30000
+                    gte = 0
+                }
+                max {
+                    lte = 30000
+                    gte = 0
+                }
+                mean {
+                    lte = 30000
+                    gte = 0
+                }
+            }
+            failedRequests {
+                percent {
+                    lte = 5
+                    gte = 0
+                }
+            }
+            successfulRequests {
+                percent {
+                    lte = 100
+                    gte = 0
+                }
+            }
+        }
+    }
+}
