+++ modules/openapi-generator/src/main/resources/kotlin-gatling/gatling.mustache
@@ -1,0 +1,23 @@
+#########################
+# Gatling Configuration #
+#########################
+
+# If a setting is not specified, the default will be used.
+# We prefer to use latest defaults, unless stated otherwise.
+# Defaults at https://github.com/gatling/gatling/blob/main/gatling-core/src/main/resources/gatling-defaults.conf
+
+gatling {
+    charting {
+        noReports = true       # When set to true, don't generate HTML reports
+    }
+    http {
+        warmUpUrl = "https://acc.ah.nl"          # The URL to use to warm-up the HTTP stack (blank means disabled)
+        enableGA = false                           # Very light Google Analytics (Gatling and Java version), please support
+    }
+    data {
+        writers = [console, graphite]      # The list of DataWriters to which Gatling write simulation data (currently supported : console, file, graphite)
+        graphite {
+            writePeriod = 10          # GraphiteDataWriter's write interval, in seconds
+        }
+    }
+}
