+++ modules/openapi-generator/src/main/resources/kotlin-gatling/httpProtocolConfiguration.mustache
@@ -1,0 +1,31 @@
+package {{ configurationPackage }}
+
+import {{ configurationPackage }}.*
+import io.gatling.javaapi.http.HttpDsl
+import io.gatling.javaapi.http.HttpDsl.http
+import io.gatling.javaapi.http.HttpProtocolBuilder
+
+
+/**
+* Definition of the httpProtocol, will be used for all requests in scenario
+*/
+object HttpProtocolConfiguration {
+
+    val httpProtocol: HttpProtocolBuilder = http
+        .baseUrl(TestConfiguration.targetBaseUrl)
+        .inferHtmlResources()
+        .acceptHeader("*/*")
+        .acceptEncodingHeader("gzip, deflate, sdch")
+        .acceptLanguageHeader("en-US,en;q=0.8")
+        .contentTypeHeader("application/json")
+        .userAgentHeader("Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/54.0.2840.98 Safari/537.36")
+        .shareConnections()
+        .header("X-Authorization", "#{jwt}")
+        .run {
+            if (TestConfiguration.useProxy) {
+                proxy(HttpDsl.Proxy("localhost", 8888))
+            } else {
+                this
+            }
+        }
+}
