+++ modules/openapi-generator/src/main/resources/kotlin-gatling/members-acc-small.csv
@@ -1,0 +1,11 @@
+memberId,email,password
+31300465,ahold.perftest+01421@mailinator.com,AHperftest2021!
+31300466,ahold.perftest+01422@mailinator.com,AHperftest2021!
+31300472,ahold.perftest+01423@mailinator.com,AHperftest2021!
+31300473,ahold.perftest+01424@mailinator.com,AHperftest2021!
+31300474,ahold.perftest+01425@mailinator.com,AHperftest2021!
+31300475,ahold.perftest+01426@mailinator.com,AHperftest2021!
+31300476,ahold.perftest+01427@mailinator.com,AHperftest2021!
+31300477,ahold.perftest+01428@mailinator.com,AHperftest2021!
+31300478,ahold.perftest+01429@mailinator.com,AHperftest2021!
+31300479,ahold.perftest+01430@mailinator.com,AHperftest2021!
