+++ modules/openapi-generator/src/main/resources/scala-gatling/configurationDumper.mustache
@@ -1,0 +1,26 @@
+package {{ helperPackage}}
+
+import java.util.Properties
+
+import {{ configurationPackage }}._
+
+object ConfigurationDumper {
+
+    val values: String = {
+
+    var values: String  = "\n"+"="*100+"\nSystem Properties values:\n\n"
+    val systemProperties : Properties = System.getProperties
+    systemProperties.forEach( (key, value) => values += "System.getProperty(\"" + key + "\") = " + value + "\n")
+    values += "="*100 + "\nTestConfiguration values:\n\n"
+
+    for (field <- TestConfiguration.getClass.getDeclaredFields()) {
+    field.setAccessible(true)
+    values += s"TestConfiguration.${field.getName} = ${field.get(TestConfiguration)}\n"
+    }
+
+    println(values) // I print this now, so it will be printed during creation of object instances etc, before running
+    values
+    }
+
+
+}
