+++ modules/openapi-generator/src/main/resources/scala-gatling/gitignore.mustache
@@ -1,0 +1,3 @@
+results
+target
+.DS_store
