+++ modules/openapi-generator/src/main/resources/scala-gatling/idePathHelper.mustache
@@ -1,0 +1,19 @@
+package {{ helperPackage}}
+
+import java.nio.file.Paths
+
+/**
+* Utility object to create the paths used throughout this loadtest script.
+*/
+object IDEPathHelper {
+
+    private val projectRootDir = Paths.get(getClass.getClassLoader.getResource("gatling.conf").toURI).getParent.getParent.getParent
+    private val mavenTargetDirectory = projectRootDir.resolve("target")
+    private val mavenSrcTestDirectory = projectRootDir.resolve("src").resolve("test")
+
+    val mavenSourcesDirectory = mavenSrcTestDirectory.resolve("scala")
+    val mavenResourcesDirectory = mavenSrcTestDirectory.resolve("resources")
+    val mavenBinariesDirectory = mavenTargetDirectory.resolve("test-classes")
+    val resultsDirectory = mavenTargetDirectory.resolve("gatling")
+    val recorderConfigFile = mavenResourcesDirectory.resolve("recorder.conf")
+}
