+++ modules/openapi-generator/src/main/resources/scala-gatling/members-acc-small.csv
@@ -1,0 +1,11 @@
+memberId
+18464431
+18465134
+18464412
+18464411
+18464424
+18464425
+18464426
+18464427
+18464428
+18464429
